Rough observable prototype
The library now contains a quick implementation of the observer pattern
using CATs.

No error repoprting is provided (some of the error messages generated
for incorrect code are really loooooong), and only minimal testing is
added for now, but it should work as long as it is used correctly.

TODO: tons of sanity checks and code cleanup.